+++ src/package/packageVersionCreate.ts
@@ -277,6 +277,7 @@ export class PackageVersionCreate {
       SourceOrg: this.options.sourceorg,
       CalculateCodeCoverage: this.options.codecoverage ?? false,
       SkipValidation: this.options.skipvalidation ?? false,
+      AsyncValidation: this.options.asyncvalidation ?? false,
       // note: the createRequest's Language corresponds to the AllPackageVersion's language
       Language: this.options.language,
     };
@@ -290,6 +291,17 @@ export class PackageVersionCreate {
         );
       }
       delete requestObject.Language;
+    }
+
+    // Ensure we only include the async validation property for a connection api version
+    // of v60.0 or higher.
+    if (this.connection.getApiVersion() <= '60.0') {
+      if (requestObject.AsyncValidation) {
+        this.logger.warn(
+          `The async validation option is only valid for API version 60.0 and higher. Ignoring ${requestObject.AsyncValidation}`
+        );
+      }
+      delete requestObject.AsyncValidation;
     }
 
     if (preserveFiles) {